addressed pr concerns

+++ ttnn/cpp/ttnn/operations/normalization/layernorm/device/kernels/dataflow/reader_unary_interleaved_ln.cpp
@@ -6,7 +6,6 @@
 #include "dataflow_api.h"
 #include "ttnn/deprecated/tt_dnn/kernels/dataflow/generate_reduce_scaler.hpp"
 #include "ttnn/deprecated/tt_dnn/kernels/dataflow/generate_bcast_scalar.hpp"
-#include "ttnn/cpp/ttnn/operations/kernel_helper_functions/pad_tile.hpp"
 
 template <bool DRAM, uint32_t tile_hw = 1024>
 void read_row_to_cb(
@@ -71,7 +70,8 @@ void kernel_main() {
     // Generate constant tiles for layernorm compute
     {
         constexpr uint32_t cb_in_2 = tt::CBIndex::c_2;
-        uint32_t scaler = (uint16_t)get_arg_val<uint32_t>(4);
+        // doubly packed float into a uint32_t = 1/n where n is the length of the dimension we are averaging;
+        uint32_t scaler = get_arg_val<uint32_t>(4);
         generate_reduce_scaler(cb_in_2, scaler);
         // doubly packed float into a uint32_t = 1.0f;
         uint32_t one_scaler = 0x3f803f80;
